# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WIT Sprint OS API
  version: "1.0.0"
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    REST API for the WIT Sprint OS delivery hierarchy.

    ## Naming

    The URL vocabulary predates the UI's, and the two do not match. Read it as:

    | URL / payload | The UI calls it |
    | ------------- | --------------- |
    | `product`     | **Module**      |
    | `module`      | **Component**   |

    The hierarchy is `client → project → product → module → sprint → task`,
    which the UI presents as Client → Project → Module → Component → Sprint.

    ## Conventions

    - **PATCH is partial.** Only the keys present in the body are written; an
      absent key leaves its column untouched. This is what makes two concurrent
      edits to different fields both survive. `PATCH /products/{id}/current-sprint`
      is the single exception where an explicit `null` differs from an absent
      key — see that operation.
    - **Ids may be supplied or generated.** Send `id` on create to choose it;
      omit it and the server mints one. A duplicate answers `409`.
    - **Lineage is enforced.** A child may only reference a parent inside its own
      product: a sprint's component, a task's backlog item, a report's sprint. A
      mismatch is a `400`, not a silent link.
    - **Timestamps** are RFC 3339 with an offset (`2026-07-23T02:35:08.59+07:00`).
...
      RFC 3339.
...

    ## Authentication

...
    it stands.

...
security: []
servers:
...
paths:
...
  /modules/{moduleId}/sprints:
    parameters:
      - $ref: '#/components/parameters/ModuleId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
    get:
      tags: [modules, sprints]
      summary: List a component's sprints
      operationId: listSprintsByModule
      responses:
        '200':
          description: A page of sprints, newest number first.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Sprint' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /modules/{moduleId}/backlog:
    parameters:
      - $ref: '#/components/parameters/ModuleId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
    get:
      tags: [modules, backlog]
      summary: List a component's backlog items
      operationId: listBacklogItemsByModule
      responses:
        '200':
          description: A page of backlog items, newest first.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BacklogItem' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ------------------------------------------------------------- sprints ---
  /products/{productId}/sprints:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [sprints]
      summary: List a product's sprints
      operationId: listSprintsByProduct
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of sprints, newest number first.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Sprint' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [sprints]
      summary: Create a sprint
      description: |
        Omit `number` and the sprint is auto-numbered: the next number for this
        product, assigned under a row lock so a burst produces a gapless
        sequence rather than a pile of conflicts. Send `number` to choose it
        yourself, and a clash is a genuine `409`.

        `moduleId` must name a component of this product.
      operationId: createSprint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SprintCreate' }
      responses:
        '201':
          description: The created sprint.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sprint' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: That number is already taken, or the row was busy — retry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /sprints/{sprintId}:
    parameters:
      - $ref: '#/components/parameters/SprintId'
    get:
      tags: [sprints]
      summary: Get a sprint
      operationId: getSprint
      responses:
        '200':
          description: The sprint.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sprint' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [sprints]
      summary: Update a sprint
      operationId: updateSprint
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SprintUpdate' }
      responses:
        '200':
          description: The updated sprint.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sprint' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [sprints]
      summary: Delete a sprint
      description: Cascades to its tasks, members and committed backlog links.
      operationId: deleteSprint
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /sprints/{sprintId}/members:
    parameters:
      - $ref: '#/components/parameters/SprintId'
    get:
      tags: [sprints, members]
      summary: List a sprint's members
      description: |
        **Not paginated, and not bounded.** A sprint team is usually small, but nothing enforces that: `?limit` is ignored and the whole set
        is returned. A load test found one sprint holding 4,898 committed items
        answering 1.3 MB to a request that asked for 5, so treat the size of this
        response as a function of your data, not of the request. Paginating these
        is a known outstanding fix.
      operationId: listSprintMembers
      responses:
        '200':
          description: The sprint's members, with the person's details joined in.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SprintMember' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /sprints/{sprintId}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/SprintId'
      - $ref: '#/components/parameters/MemberId'
    put:
      tags: [sprints, members]
      summary: Add or update a sprint member
      description: Idempotent — repeating it updates the allocation.
      operationId: addSprintMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allocation:
                  type: integer
                  minimum: 0
                  maximum: 100
                  default: 100
                capacityDays:
                  type: number
                  minimum: 0
                  maximum: 999.99
                  default: 0
      responses:
        '200':
          description: The membership.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SprintMemberLink' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [sprints, members]
      summary: Remove a sprint member
      operationId: removeSprintMember
      responses:
        '204': { description: Removed. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /sprints/{sprintId}/backlog:
    parameters:
      - $ref: '#/components/parameters/SprintId'
    get:
      tags: [sprints, backlog]
      summary: List a sprint's committed backlog
      description: |
        **Not paginated, and not bounded.** `?limit` is ignored and the whole set
        is returned. A load test found one sprint holding 4,898 committed items
        answering 1.3 MB to a request that asked for 5, so treat the size of this
        response as a function of your data, not of the request. Paginating these
        is a known outstanding fix.

        Ordered by `position`.
      operationId: listSprintBacklogItems
      responses:
        '200':
          description: The committed items, with the backlog item's content joined in.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SprintBacklogItem' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /sprints/{sprintId}/backlog/{itemId}:
    parameters:
      - $ref: '#/components/parameters/SprintId'
      - $ref: '#/components/parameters/ItemId'
    put:
      tags: [sprints, backlog]
      summary: Commit a backlog item to a sprint
      description: |
        The item must belong to the same product as the sprint. Omit `position`
        and it is appended under a row lock.
      operationId: addSprintBacklogItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: integer
                  minimum: 0
                  description: Omit to append.
      responses:
        '200':
          description: The link.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SprintBacklogLink' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400':
          description: The item does not exist or belongs to a different module.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [sprints, backlog]
      summary: Remove an item from a sprint
      operationId: removeSprintBacklogItem
      responses:
        '204': { description: Removed. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ------------------------------------------------------------- backlog ---
  /products/{productId}/backlog:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [backlog]
      summary: List a product's backlog
      operationId: listBacklogItemsByProduct
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of backlog items, newest first.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BacklogItem' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [backlog]
      summary: Create a backlog item
      operationId: createBacklogItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BacklogItemCreate' }
      responses:
        '201':
          description: The created item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BacklogItem' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /backlog/{itemId}:
    parameters:
      - $ref: '#/components/parameters/ItemId'
    get:
      tags: [backlog]
      summary: Get a backlog item
      operationId: getBacklogItem
      responses:
        '200':
          description: The item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BacklogItem' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [backlog]
      summary: Update a backlog item
      operationId: updateBacklogItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BacklogItemUpdate' }
      responses:
        '200':
          description: The updated item.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BacklogItem' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [backlog]
      summary: Delete a backlog item
      description: Cascades to its tasks and to any sprint commitment of it.
      operationId: deleteBacklogItem
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # --------------------------------------------------------------- tasks ---
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Clamped to 1000; anything unparseable falls back to the default.
      schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
    Offset:
      name: offset
      in: query
      description: |
        Rows to skip. Postgres walks and discards them, so deep offsets get
        progressively more expensive — prefer narrowing by parent id. Values
        above 1,000,000 are refused with a 400 rather than silently answered
        with a different page.
      schema: { type: integer, minimum: 0, maximum: 1000000, default: 0 }
...
    ProductId:
      name: productId
      in: path
      required: true
      description: 'A product id — the UI calls this a Module.'
      schema: { type: string }
    ModuleId:
      name: moduleId
      in: path
      required: true
      description: 'A module id — the UI calls this a Component.'
      schema: { type: string }
    SprintId: { name: sprintId, in: path, required: true, schema: { type: string } }
...
  headers:
    XHasMore:
      description: Present and `true` when another page exists.
      schema: { type: string, enum: ['true'] }
  responses:
    BadRequest:
      description: The request is malformed, or a value is out of range.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooLarge:
      description: The body is over 1 MB.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Retryable. Either the request took too long — the statement timeout is
        5 s and the whole request is capped at 15 s — or the database could not
        take it: out of connection slots, restarting, or the connection dropped.
        Never a bug on the caller's side, and always worth retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Sprint:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        moduleId: { type: [string, 'null'], description: The component that owns this sprint. }
        number: { type: integer, minimum: 1, maximum: 100000 }
        name: { type: string }
        goal: { type: string }
        startDate: { $ref: '#/components/schemas/DateOnly' }
        endDate: { $ref: '#/components/schemas/DateOnly' }
        workingDays: { type: integer, minimum: 0 }
        daysLeft: { type: integer, minimum: 0 }
        status: { $ref: '#/components/schemas/SprintStatus' }
        committed: { type: integer, minimum: 0 }
        completed: { type: integer, minimum: 0 }
        progress: { type: integer, minimum: 0, maximum: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    SprintCreate:
      type: object
      required: [name]
      properties:
        id: { type: string }
        moduleId: { type: string, description: A component of this product. }
        number:
          type: integer
          minimum: 1
          maximum: 100000
          description: Omit to have the next number assigned for you.
        name: { type: string }
        goal: { type: string }
        startDate: { $ref: '#/components/schemas/DateOnly' }
        endDate: { $ref: '#/components/schemas/DateOnly' }
        workingDays: { type: integer, minimum: 0, default: 10 }
        daysLeft: { type: integer, minimum: 0, default: 10 }
        status: { $ref: '#/components/schemas/SprintStatus' }
        committed: { type: integer, minimum: 0 }
        completed: { type: integer, minimum: 0 }
        progress: { type: integer, minimum: 0, maximum: 100 }
        risk: { $ref: '#/components/schemas/Risk' }
...
    BacklogItem:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        moduleId: { type: [string, 'null'] }
        title: { type: string }
        story: { type: string }
        acceptanceCriteria: { type: array, items: { type: string } }
        type:
          type: string
          description: 'Free text; the UI offers story, bug, spike, chore.'
        priority:
          type: string
          description: 'Free text; the UI offers low, medium, high, critical.'
        readiness: { $ref: '#/components/schemas/Readiness' }
        estimate: { type: integer, minimum: 0 }
        aiSuggestions: { type: array, items: { type: string } }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...